graphql module version update for js/node_modules
Summary:This bumps graphql from 0.4.13 to 0.4.18. It's a simple copy from https://github.com/graphql/graphql-js.git to {FBReactKit/catalyst}/js/node_modules

The reason is that babel-plugin-relay-native-fragments requires the latest version.

The babel-relay-plugin package.json manifest says it depends on 0.4.17 of graphql. However the actual graphql module is at 0.4.13. Unless babel-relay-plugin needs to be on an earlier version of the graphql module it seems sensible to update it.

Reviewed By: martinbigio

Differential Revision: D2994989

fb-gh-sync-id: 6f97341cc2157583b3730a363d024f8990ea3b28
shipit-source-id: 6f97341cc2157583b3730a363d024f8990ea3b28