Make plugin build checker independent of Babel version
Summary:Currently, the build checker test is tied to the exact output of the transform.
This means if people run different minor versions of babel, the test fails.

This change makes the test more resilient. Everytime the build is run, the
`SRC_HASH` is updated. The test hashes the source files again and makes sure the
hash matches. This makes sure we ran the build after modifying some sources.
Closes https://github.com/facebook/relay/pull/901

Reviewed By: yungsters

Differential Revision: D3010038

Pulled By: kassens

fb-gh-sync-id: 45eb2c82333b470d2c6151d5cdf82352c5042d10
shipit-source-id: 45eb2c82333b470d2c6151d5cdf82352c5042d10